fix: reset template
